feat(map): implement user carpool states
This commit introduces the management of user carpool states within the map view.

It adds an enum `EUserCarpoolState` to represent the different states a user can be in during the carpool process (SEARCHING, WAITING_FOR_CARPOOL_START, IN_CARPOOL, COMPLETED, CANCELLED).

The map view now observes the user's carpool state and displays different UI components accordingly. For example, when the user is in the `WAITING_FOR_CARPOOL_START` state, the `WaitForCarpoolStart` view is shown..